feat(note): добавлена функциональность связанных заметок
- Добавлено поле `RelatedIDs` в модель `Note` для хранения ID связанных заметок
- Обновлена структура `CreateNoteInput` для включения `RelatedIDs` при создании заметок со связями
- Модифицирован `CreateNoteHandler` для обработки `RelatedIDs` при создании заметки
- Обновлена Swagger-документация для отражения изменений в модели заметки и входных данных
- Добавлены проверки существования и принадлежности связанных заметок пользователю
- Реализована дедупликация ID связанных заметок

+++ internal/response/response.go
@@ -41,7 +41,7 @@ type NoteResponse struct {
 	Attachments []AttachmentShort `json:"attachments,omitempty"`
 	IsArchived  bool              `json:"is_archived"`
 	Tags        []TagShort        `json:"tags,omitempty"`
-	RelatedIDs  []uint            `json:"related_ids,omitempty"`
+	RelatedIDs  []int64           `json:"related_ids,omitempty"`
 	CreatedAt   string            `json:"created_at"`
 	UpdatedAt   string            `json:"updated_at"`
 }
